Update README.md to use relative links
Updated the README to use relative links to this repo's examples, files, and folders. This all started when I clicked on the "complicated search forms" example (example/app/models/post_search.rb) and noticed that the link was broken. If you would prefer to not use relative links, I can just fix this link instead, I believe the rest of the links were correctly working.